+++ packages/blade/src/components/SpotlightPopoverTour/TourPopover.native.tsx
@@ -1,0 +1,14 @@
+/* eslint-disable react/jsx-no-useless-fragment */
+/* eslint-disable @typescript-eslint/no-explicit-any */
+import { throwBladeError } from '~utils/logger';
+
+const TourPopover = (_: any): React.ReactElement => {
+  throwBladeError({
+    message: 'Tour is not yet implemented for native',
+    moduleName: 'Tour',
+  });
+
+  return <></>;
+};
+
+export { TourPopover };