+++ packages/blade/src/utils/storybook/StoryPageWrapper.tsx
@@ -3,7 +3,7 @@ import { ArgsTable, Primary, PRIMARY_STORY, Stories } from '@storybook/addon-doc
 import styled from 'styled-components';
 import useMakeFigmaURL from './useMakeFigmaURL';
 import FigmaEmbed from './FigmaEmbed';
-import { SandboxHighlighter } from './Sandbox';
+import { SandboxHighlighter } from './Sandbox/SandpackEditor';
 import { componentData } from './componentStatusData';
 import BaseBox from '~components/Box/BaseBox';
 import { Alert } from '~components/Alert';